requirements.txt
Django==3.1
django-cors-headers==3.8.0
django-environ==0.4.5
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-camel-case==1.2.0
Markdown==3.3.4
